Fix StringExtension registration
