+++ src/SonataMediaBundle.php
@@ -17,6 +17,7 @@ use Sonata\CoreBundle\Form\FormHelper;
 use Sonata\MediaBundle\DependencyInjection\Compiler\AddProviderCompilerPass;
 use Sonata\MediaBundle\DependencyInjection\Compiler\GlobalVariablesCompilerPass;
 use Sonata\MediaBundle\DependencyInjection\Compiler\ThumbnailCompilerPass;
+use Sonata\MediaBundle\DependencyInjection\Compiler\TwigStringExtensionCompilerPass;
 use Sonata\MediaBundle\Form\Type\ApiDoctrineMediaType;
 use Sonata\MediaBundle\Form\Type\ApiGalleryHasMediaType;
 use Sonata\MediaBundle\Form\Type\ApiGalleryType;
@@ -35,6 +36,7 @@ class SonataMediaBundle extends Bundle
         $container->addCompilerPass(new AddProviderCompilerPass());
         $container->addCompilerPass(new GlobalVariablesCompilerPass());
         $container->addCompilerPass(new ThumbnailCompilerPass());
+        $container->addCompilerPass(new TwigStringExtensionCompilerPass());
 
         $this->registerFormMapping();
     }
