+++ tests/DependencyInjection/Compiler/AddProviderCompilerPassTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\DependencyInjection\Tests\Compiler;
+namespace Sonata\MediaBundle\Tests\DependencyInjection\Compiler;
 
 use PHPUnit\Framework\TestCase;
 use Sonata\MediaBundle\DependencyInjection\Compiler\AddProviderCompilerPass;
